+++ src/apis/message.ts
@@ -67,3 +67,19 @@ export const postMessage = async (message: WriteMessageType) => {
     return false
   }
 }
+
+export const getMyMessages = async () => {
+  try {
+    const {
+      data: {
+        data: { myLetterInfos },
+      },
+    } = await client.get('/letter')
+    console.log(myLetterInfos)
+    return myLetterInfos
+  } catch (error) {
+    if (error instanceof Error) {
+      console.error(error.message)
+    }
+  }
+}
